+++ backend/core/exceptions.py
@@ -1,0 +1,22 @@
+class EventIsFull(Exception):
+    pass
+
+
+class NotBookable(Exception):
+    pass
+
+
+class UserEventNbExceded(Exception):
+    pass
+
+
+class UserEventNbExcededAdmin(Exception):
+    pass
+
+
+class UserEventNbExcededUser(Exception):
+    pass
+
+
+class ParticipantNbExceded(Exception):
+    pass