+++ backend/test/test_api_user.py
@@ -8,7 +8,7 @@ from sqlalchemy import select
 
 from core.models import GTEvents, TReservations
 from core.models import TTokens
-from core.routes import EventIsFull
+from core.exceptions import EventIsFull
 from core.env import db
 
 from .utils import login
@@ -36,6 +36,34 @@ TEST_RESERVATION = {
     "nb_6_8_ans": 1,
     "nb_9_12_ans": 2,
     "nb_plus_12_ans": 2,
+    "num_departement": "48",
+}
+
+TEST_RESERVATION_1_PERSONNE = {
+    "nom": "YO",
+    "prenom": "yo",
+    "commentaire": "saisie test",
+    "tel": "00 00 00 00 00 ",
+    "email": "user.tropresa@test.fr",
+    "nb_adultes": 1,
+    "nb_moins_6_ans": 0,
+    "nb_6_8_ans": 0,
+    "nb_9_12_ans": 0,
+    "nb_plus_12_ans": 0,
+    "num_departement": "48",
+}
+
+TEST_RESERVATION_PLEIN_PERSONNES = {
+    "nom": "YO",
+    "prenom": "yo",
+    "commentaire": "saisie test",
+    "tel": "00 00 00 00 00 ",
+    "email": "user.tropmonde@test.fr",
+    "nb_adultes": 10,
+    "nb_moins_6_ans": 10,
+    "nb_6_8_ans": 10,
+    "nb_9_12_ans": 10,
+    "nb_plus_12_ans": 0,
     "num_departement": "48",
 }
 
@@ -94,3 +122,65 @@ class TestAPI:
         response = self.client.get(url_for("app_routes.get_reservations"))
 
         assert response.status_code == 200
+
+    def test_post_limit_nb_animations(self, events):
+        login(self.client, "user@test.fr")
+        # Create reservation
+        event = db.session.scalars(
+            select(GTEvents)
+            .where(GTEvents.name == "Pytest bookable")
+            .order_by(GTEvents.id.desc())
+        ).first()
+
+        data_resa = TEST_RESERVATION_1_PERSONNE
+        data_resa["id_event"] = event.id
+
+        nb_limit_per_user = current_app.config["NB_ANIM_MAX_PER_USER"]
+
+        # Création du nombre de reservations spécifié dans NB_ANIM_MAX_PER_USER
+        for loop in range(nb_limit_per_user):
+            resp = post_json(
+                self.client, url_for("app_routes.post_reservations"), data_resa
+            )
+            assert resp == 200
+            # Confirm resa
+            data = json.loads(resp.data)
+            id_resa = data["id_reservation"]
+            resa = db.session.get(TReservations, id_resa)
+            resp = post_json(
+                self.client,
+                url_for("app_routes.confirm_reservation"),
+                {"resa_token": resa.token},
+            )
+            assert resp == 200
+
+        #  Ajout de 1 reservation ce qui doit retourner une erreur
+        resp = post_json(
+            self.client, url_for("app_routes.post_reservations"), data_resa
+        )
+        assert resp.status_code == 422
+        assert (
+            json_of_response(resp)["error"]
+            == "Vous avez atteint la limite du nombre de réservations possible par personne"
+        )
+
+    def test_post_limit_nb_participants(self, events):
+        login(self.client, "user@test.fr")
+        # Create reservation
+        event = db.session.scalars(
+            select(GTEvents)
+            .where(GTEvents.name == "Pytest bookable")
+            .order_by(GTEvents.id.desc())
+        ).first()
+
+        data_resa = TEST_RESERVATION_PLEIN_PERSONNES
+        data_resa["id_event"] = event.id
+
+        resp = post_json(
+            self.client, url_for("app_routes.post_reservations"), data_resa
+        )
+        assert resp.status_code == 422
+        assert (
+            json_of_response(resp)["error"]
+            == "Vous ne pouvez pas inscrire autant de personnes sur une animation"
+        )